AP-13721: Fixed bug in createLocation, so it also works for LOCAL FS
The createLocation method checked, if the file system specifier is null
and threw an exception if so. Because of this behavior the methode
didn't work for local FS, because the file system specifier is always
null in this case. Now it just passes null to FSLocation constructor
instead of throwing an exception.

AP-13721 (String to Path node)