AP-13432: Make displayed convenience file system configurable
So far Custom/KNIME URL was excluded if File was not among the possible
selection modes. This was enforced in the constructor of
AbstractDialogComponentFileChooser which is a bad place for such a
constraint because this kind of constrained heavily depends on the node
functionality. Therefore this commit allows programmers to specify which
convenience file systems should be available in the dialog by passing a
set of FSCategory when creating the settings model for the file chooser.

AP-13432 (Delete files no table input (new file handling))