AP-16176: Fix bug when enforcing types for columns that no longer have a type
Up until now we tried to find an alternative path to the same output
type and failed if no such path exists. However, it would be better to
use the configured production path because it is what the user decided
on and we don't have any information to the contrary.

AP-16176 (Decouple Serialization and UI from Type Mapping Framework)