AP-14001: Remove checkAccess() implementation from Local Relative To
Reason: checkAccess() is unsupported for the Relative to filesystems.
in general, because it cannot be implemented for the server-side (REST-based)
Relative to filesystems (the KNIME Server REST API does not provide it).
Hence we need to remove it here to have consistent behavior.

AP-14001 (Filehandling Refactoring: Create extension point for NIO File System implementations)