AP-20714: Enhance FSPath to resolve to local file if possible
In order to work with Path instances, it is sometimes necessary to make sure that the resulting file is on the
local file system, e.g. when a library cannot (reliably) work with an InputStream and needs random access to the file.
In cases where the Path already denotes a file on the local file system, it would be wasteful to copy that file to the
temp directory.
Therefore, all Path implementations that could describe a local file are enhanced to return a "local path".

The default implementation ensures existing classes extending FSPath do not break.

AP-20714 (POI Update: Research file versus stream performance)